primitives with And-gate plus small demo
This commit includes a few primitives such as

* a Wire, which state can be On or Off;
* two special wires, Power and Ground, that are always On and Off;
* a LED, which emit on stdout everytime its state changes;
* a Diode, with an input wire and an output wire, which can be used to
  both isolate components and exposes easy-to-plug wires;
* a Transistor, with its three wires.

In addition to those primitives, we also included a And-gate which uses
two transistors:

    power --> transistor --> transistor --> output
                  ^              ^
                  |              |
               input 1        input 2

The general mechanism is that components are plugged in order, power and
input 1 then transitor then input 2 then transistor then output. Every
time the state of a component changes (e.g. when we switch on the first
input) then this component *signals* the next component. The next
component then reads the state of the previous component, updates its
state accordingly and optionnaly signals the next component.